9C-1141-6 Api Module. Account Package
We refactor the Api Module. We extract an `account` Api module,
for the `/users` and `/installations` routes.
