+++ modules/api/src/main/scala/cards/nine/api/JsonFormats.scala
@@ -1,10 +1,7 @@
 package cards.nine.api
 
 import cards.nine.api.messages.GooglePlayMessages._
-import cards.nine.api.messages.InstallationsMessages._
-import cards.nine.api.messages.UserMessages._
 import cards.nine.domain.application.{ Package, Widget }
-import cards.nine.domain.account._
 import spray.httpx.SprayJsonSupport
 import spray.json._
 
@@ -16,44 +13,6 @@ trait JsonFormats
     def read(json: JsValue): Package = Package(StringJsonFormat.read(json))
     def write(pack: Package): JsValue = StringJsonFormat.write(pack.value)
   }
-
-  implicit object EmailJsonFormat extends JsonFormat[Email] {
-    def read(json: JsValue): Email = Email(StringJsonFormat.read(json))
-    def write(pack: Email): JsValue = StringJsonFormat.write(pack.value)
-  }
-
-  implicit object DeviceTokenJsonFormat extends JsonFormat[DeviceToken] {
-    def read(json: JsValue): DeviceToken = DeviceToken(StringJsonFormat.read(json))
-    def write(pack: DeviceToken): JsValue = StringJsonFormat.write(pack.value)
-  }
-
-  implicit object ApiKeyJsonFormat extends JsonFormat[ApiKey] {
-    def read(json: JsValue): ApiKey = ApiKey(StringJsonFormat.read(json))
-    def write(pack: ApiKey): JsValue = StringJsonFormat.write(pack.value)
-  }
-
-  implicit object AndroidIdJsonFormat extends JsonFormat[AndroidId] {
-    def read(json: JsValue): AndroidId = AndroidId(StringJsonFormat.read(json))
-    def write(pack: AndroidId): JsValue = StringJsonFormat.write(pack.value)
-  }
-
-  implicit object GoogleIdTokenJsonFormat extends JsonFormat[GoogleIdToken] {
-    def read(json: JsValue): GoogleIdToken = GoogleIdToken(StringJsonFormat.read(json))
-    def write(pack: GoogleIdToken): JsValue = StringJsonFormat.write(pack.value)
-  }
-
-  implicit object SessionTokenJsonFormat extends JsonFormat[SessionToken] {
-    def read(json: JsValue): SessionToken = SessionToken(StringJsonFormat.read(json))
-    def write(pack: SessionToken): JsValue = StringJsonFormat.write(pack.value)
-  }
-
-  implicit val apiLoginRequestFormat = jsonFormat3(ApiLoginRequest)
-
-  implicit val apiLoginResponseFormat = jsonFormat2(ApiLoginResponse)
-
-  implicit val updateInstallationRequestFormat = jsonFormat1(ApiUpdateInstallationRequest)
-
-  implicit val updateInstallationResponseFormat = jsonFormat2(ApiUpdateInstallationResponse)
 
   implicit val apiGetRecommendationsByCategoryRequestFormat = jsonFormat2(ApiGetRecommendationsByCategoryRequest)
