--- modules/api/src/main/scala/cards/nine/api/messages/InstallationsMessages.scala
@@ -1,10 +1,0 @@
-package cards.nine.api.messages
-
-import cards.nine.domain.account.{ AndroidId, DeviceToken }
-
-object InstallationsMessages {
-
-  case class ApiUpdateInstallationRequest(deviceToken: Option[DeviceToken])
-
-  case class ApiUpdateInstallationResponse(androidId: AndroidId, deviceToken: Option[DeviceToken])
-}
